requirements.txt
streamlit==1.33.0
tensorflow==2.13.0
scikit-learn
Pillow
numpy
matplotlib
